New UML Replaced JPG (exported from an iPad app) with UML. There are other changes.
Document ID is now MD5. This is intended to be the MD5 sum of the body.

Person now carries authentication data. E-mail is intended to be the username; passhash is intended to store whatever the server needs for evaluating subsequent password entries.

Person now carries a UUID. The purpose of a UUID is to have a stable identifier even if data is replicated between databases.

The reason for dropping Auth0 is to reduce the total number of dependencies, and particularly to eliminate institutions (such as Auth0) as dependencies.

For the same reason, we're sticking with keeping the document body in the relational database. This is to avoid additional dependencies (such as an object database) as well as to avoid institutions such as Amazon (via S3).

Whereas the documents are immutable, document pages are very stable. These can therefore be cached in the web server (e.g. Nginx).